:sparkles: feat(backend): mention-position에 따른 sort설정 (#277)

+++ backend/src/models/entities/mention.entity.ts
@@ -67,7 +67,7 @@ export class MentionEntity extends DefaultEntity {
 	@IsUUID(4, { message: uuidValidationMessage })
 	public readonly mentionTypeId!: string;
 
-	@Column('numeric', { name: 'mentionPosition', default: 0 })
+	@Column('int4', { name: 'mentionPosition', default: 0 })
 	@ApiProperty({
 		nullable: false,
 	})
